requirements.txt
pyyaml
# parse